+++ packages/notification/src/ToastNotification.spec.tsx
@@ -1,4 +1,4 @@
-import { render, screen, waitFor } from "@testing-library/react"
+import { render, screen } from "@testing-library/react"
 import * as React from "react"
 import ToastNotification from "./ToastNotification"
 
@@ -17,7 +17,5 @@ test("The basic notification renders correctly", async () => {
     </ToastNotification>
   )
 
-  await waitFor(async () =>
-    expect(screen.getByRole("heading").textContent).toEqual("Success")
-  )
+  await screen.findByText("Success")
 })